SRE-697 reorganize environments app
